Flyttet endepunkt for liste av sykmeldinger til SykmeldingServiceuserApi
